Use 'unsigned long' for timer_get_tics
This is what is expected by jiffies.h
(On most 32b machines, uint32_t === unsigned long, but not on 64b hosts)

+++ softpll/softpll_ng.h
@@ -184,7 +184,7 @@ struct spll_aux_state {
 struct softpll_state {
 	int mode;
 	int seq_state;
-	uint32_t dac_timeout;
+	unsigned long dac_timeout;
 	int delock_count;
 	unsigned irq_count;
 	unsigned ref_count, tag_count;
